Added capability to change the location of the epitope
The code only considered the first index of the epitope tag for the location, added the option to now use the middle of the epitope for the intensity vector, or the back of the epitope.

for example:

with tag DDYDDK:

NNNN DDYDDK NNNN

front: tag_epitopes = [4]

back: tag_epitopes = [10]

middle: tag_epitopes = [7]

This effects the intensity vector by changing the probe vector by an offset.