Fixed a bug in the blank model source c++ file
Custom models would fail when reaction propensities would go to a sum of 0, for example when harringtonine is applied and there are no fast/state reactions to keep the SSA going without throwing a divide by zero error. As a fix the sum of propensities is checked, if its 0 the code will set current time to 1 + final time, then fill all arrays, and return correctly.